Rename cloud_testing to testing_util.
Thanks to @mbrukman for the suggestion, the other name was kind
of repetitive.
