+++ google/cloud/google_cloud_cpp_testing.bzl
@@ -1,10 +1,10 @@
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 google_cloud_cpp_testing_HDRS = [
-    "cloud_testing/check_predicate_becomes_false.h",
-    "cloud_testing/environment_variable_restore.h",
+    "testing_util/check_predicate_becomes_false.h",
+    "testing_util/environment_variable_restore.h",
 ]
 
 google_cloud_cpp_testing_SRCS = [
-    "cloud_testing/environment_variable_restore.cc",
+    "testing_util/environment_variable_restore.cc",
 ]
