+++ google/cloud/testing_util/environment_variable_restore.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_CLOUD_TESTING_ENVIRONMENT_VARIABLE_RESTORE_H_
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_CLOUD_TESTING_ENVIRONMENT_VARIABLE_RESTORE_H_
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_TESTING_UTIL_ENVIRONMENT_VARIABLE_RESTORE_H_
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_TESTING_UTIL_ENVIRONMENT_VARIABLE_RESTORE_H_
 
 #include "google/cloud/version.h"
 #include <cstdlib>
@@ -21,7 +21,7 @@
 namespace google {
 namespace cloud {
 inline namespace GOOGLE_CLOUD_CPP_NS {
-namespace cloud_testing {
+namespace testing_util {
 /**
  * Helper class to restore the value of environment variables.
  *
@@ -45,9 +45,9 @@ class EnvironmentVariableRestore {
   char const* previous_;
 };
 
-}  // namespace cloud_testing
+}  // namespace testing_util
 }  // namespace GOOGLE_CLOUD_CPP_NS
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_CLOUD_TESTING_ENVIRONMENT_VARIABLE_RESTORE_H_
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_TESTING_UTIL_ENVIRONMENT_VARIABLE_RESTORE_H_
